Just general refactor
