+++ Cargo.toml
@@ -23,7 +23,7 @@ panic = "abort"
 
 [features]
 default = ["hvm_cli_options"]
-hvm_cli_options = []
+hvm_cli_options = ["dep:proc-macro2", "dep:quote", "dep:rust-format"]
 
 [dependencies]
 proc-macro2 = "1.0.69"
