+++ src/lib.rs
@@ -5,10 +5,7 @@
 #![allow(non_upper_case_globals)]
 
 pub mod ast;
-pub mod codegen;
 pub mod fns;
-pub mod ir;
 pub mod jit;
-pub mod quoting;
 pub mod run;
 pub mod runtime;
